Manual

The PSRClassesInterface module provides interfaces to access data structured by PSR to be used in its models. Currently there are two main interfaces.

The interface for studies. This interface is designed to read parameters from the files, some examples are deficit costs, fuel costs, currency, storage capacity etc.

The interface for reading and writing time series data. Time series data in the context of most studies have 4 dimensions (agents, stages, scenarios and blocks). Since studies of renewables with multiple agents, scenarios and stages can get quite big, we have designed different formats that are optimized to some objective (human readability, size, fast reading and writing, etc.).

Both interfaces are defined as a set of methods that need to be implemented to make a different file format work. In this manual we will describe the abstract methods and give concrete examples of code to perform the work needed.

When using the PSRClassesInterface package in your codebase we strongly advise you to create a constant PSRI to keep the code concise and explicitly declare that a certain function came from PSRClassesInterface. This can be done by adding the following code to the top of the code

using PSRClassesInterface
const PSRI = PSRClassesInterface

Abstract study interface

PSRClassesInterface.AbstractStudyInterface — Type

AbstractStudyInterface

PSRClassesInterface.RelationType — Type

RelationType

Possible relation types used in mapping function such as get_map, get_reverse_map, etc.

The current possible relation types are:

RELATION_1_TO_1
RELATION_1_TO_N
RELATION_FROM
RELATION_TO
RELATION_TURBINE_TO
RELATION_SPILL_TO
RELATION_INFILTRATE_TO
RELATION_STORED_ENERGY_DONWSTREAM
RELATION_BACKED

PSRClassesInterface.StageType — Type

PSRI.StageType

Possible stage types used in for reading and writing time series files.

The current possible stage types are:

STAGE_WEEK
STAGE_MONTH
STAGE_DAY
STAGE_YEAR

PSRClassesInterface.initialize_study — Function

initialize_study(::AbstractStudyInterface; kwargs...)

Initialize all data structures of the study.

Differences between the OpenInterface and ClassicInterface

Each study interface has its own set of kwargs... The easiest way to inspect the current available options is searching for this function on the Github repo of the desired interface.

Example:

data = PSRI.initialize_study(
    PSRI.OpenInterface(),
    data_path = PATH_CASE_EXAMPLE_BATTERIES
)

PSRClassesInterface.get_vector — Function

get_vector

PSRClassesInterface.max_elements — Function

max_elements(data::AbstractData, collection::String)

Returns an Int32 with the maximum number of elements for a given collection.

Example:

PSRI.max_elements(data, "PSRThermalPlant")

PSRClassesInterface.get_map — Function

get_map(
    data::AbstractData,
    lst_from::String,
    lst_to::String;
    allow_empty::Bool = true,
    relation_type::RelationType = RELATION_1_TO_1, # type of the direct relation
)

Returns a Vector{Int32} with the map between collections given a certain RelationType.

Examples:

PSRI.get_map(data, "PSRBattery", "PSRSystem")
PSRI.get_map(data, "PSRMaintenanceData", "PSRThermalPlant")

PSRI.get_map(data, "PSRHydroPlant", "PSRHydroPlant", relation_type = PSRI.RELATION_TURBINE_TO)
PSRI.get_map(data, "PSRHydroPlant", "PSRHydroPlant", relation_type = PSRI.RELATION_SPILL_TO)
PSRI.get_map(data, "PSRHydroPlant", "PSRHydroPlant", relation_type = PSRI.RELATION_INFILTRATE_TO)
PSRI.get_map(data, "PSRHydroPlant", "PSRHydroPlant", relation_type = PSRI.RELATION_STORED_ENERGY_DONWSTREAM)

@test PSRI.get_map(data, "PSRInterconnection", "PSRSystem", relation_type = PSRI.RELATION_FROM)
@test PSRI.get_map(data, "PSRInterconnection", "PSRSystem", relation_type = PSRI.RELATION_TO)

PSRClassesInterface.get_vector_map — Function

get_vector_map

PSRClassesInterface.get_reverse_map — Function

get_reverse_map

PSRClassesInterface.get_reverse_vector_map — Function

get_reverse_vector_map

PSRClassesInterface.get_parms — Function

get_parms(
    data::AbstractData,
    col::String,
    name::String,
    ::Type{T};
    check_type::Bool = true,
    check_parm::Bool = true,
    ignore::Bool = false,
    default::T = _default_value(T),
) where T

Returns a Vector{T} containing the elements in col to a vector in julia. This function is used to get data from collections that don't vary over time

Example:

PSRI.get_parms(data, "PSRBattery", "Einic", Float64)
PSRI.get_parms(data, "PSRBattery", "ChargeRamp", Float64)
PSRI.get_parms(data, "PSRBattery", "DischargeRamp", Float64)

PSRClassesInterface.get_code — Function

get_code(data::AbstractData, collection::String)

Returns a Vector{Int32} containing the code of each element in collection.

Example:

PSRI.get_code(data, "PSRThermalPlant")

PSRClassesInterface.get_name — Function

get_name(data::AbstractData, collection::String)

Returns a Vector{String} containing the name of each element in collection.

Example:

PSRI.get_name(data, "PSRThermalPlant")
PSRI.get_name(data, "PSRGaugingStation")

PSRClassesInterface.mapped_vector — Function

mapped_vector(
    data::AbstractData,
    col::String,
    name::String,
    ::Type{T},
    dim1::String="",
    dim2::String="";
    ignore::Bool=false,
    map_key = col, # reference for PSRMap pointer, if empty use class name
    filters = String[], # for calling just within a subset instead of the full call
) where T

Maps a Vector{T} containing the elements in col to a vector in julia. When the function update_vectors! is called the elements of the vector will be updated to the according elements registered at the current data.time_controller.

Example:

existing = PSRI.mapped_vector(data, "PSRThermalPlant", "Existing", Int32)
pot_inst = PSRI.mapped_vector(data, "PSRThermalPlant", "PotInst", Float64)

For more information please read the example Reading basic thermal generator parameters

Differences between the OpenInterface and ClassicInterface

When using mapped_vector in the OpenInterface mode the vector will be mapped with the correct values at first hand. When using mapped_vector in the ClassicInterface mode you should call update_vectors! to get the good values for the collection, otherwise you might only get a vector of zeros.

PSRClassesInterface.go_to_stage — Function

go_to_stage(data::AbstractData, stage::Integer)

Goes to the stage in the data time controller.

PSRClassesInterface.go_to_dimension — Function

go_to_dimension

PSRClassesInterface.update_vectors! — Function

update_vectors!(data::AbstractData)

Update all mapped vectors according to the time controller inside data.

update_vectors!(data::Data, filters::Vector{String})

Update filtered classes of mapped vectors according to the time controller inside data.

PSRClassesInterface.description — Function

description

PSRClassesInterface.total_stages — Function

total_stages(data::AbstractData)

Returns the total number of stages of the case.

Example:

PSRI.total_stages(data)

PSRClassesInterface.total_scenarios — Function

total_scenarios(data::AbstractData)

Returns the total number of scenarios of the case.

Example:

PSRI.total_scenarios(data)

PSRClassesInterface.total_blocks — Function

total_blocks(data::AbstractData)

Returns the total number of blocks of the case.

Example:

PSRI.total_blocks(data)

PSRClassesInterface.total_openings — Function

total_openings(data::AbstractData)

Returns the total number of openings of the case.

Example:

PSRI.total_openings(data)

PSRClassesInterface.total_stages_per_year — Function

total_stages_per_year(data::AbstractData)

Returns the total number of stages per year of the case.

Example:

PSRI.total_stages_per_year(data)

PSRClassesInterface.stage_duration — Function

stage_duration(data::AbstractData, t::Int = data.controller_stage)

Returns the duration, in hours, of the stage t.

stage_duration(data::AbstractData, date::Dates.Date)

Returns the duration, in hours, at the stage corresponding to date.

PSRClassesInterface.block_duration — Function

block_duration(data::AbstractData, date::Dates.Date, b::Int)

Returns the duration, in hours, of the block b at the stage corresponding to date.

block_duration(data::AbstractData, t::Int, b::Int)

Returns the duration, in hours, of the block b at stage t.

block_duration(data::AbstractData, b::Int)

Returns the duration, in hours, of the block b at the current stage, set by go_to_stage.

PSRClassesInterface.block_from_stage_hour — Function

block_from_stage_hour

PSRClassesInterface.get_nonempty_vector — Function

get_nonempty_vector

Abstract reader and writer interface

PSRClassesInterface.AbstractReader — Type

PSRI.AbstractReader

PSRClassesInterface.AbstractWriter — Type

PSRI.AbstractWriter

PSRClassesInterface.file_to_array — Function

PSRI.file_to_array(::Type{T}, path::String; use_header::Bool = true, header::Vector{String} = String[]) where T <: AbstractReader

Write a file to an array

PSRClassesInterface.file_to_array_and_header — Function

PSRI.file_to_array_and_header(::Type{T}, path::String; use_header::Bool = true, header::Vector{String} = String[]) where T <: AbstractReader

Write a file to an array and header

PSRClassesInterface.open — Function

PSRI.open(::Type{<:AbstractWriter}, path::String; kwargs...)

Method for opening file and registering time series data. If specified file doesn't exist, the method will create it, otherwise, the previous one will be overwritten. Returns updated AbstractWriter instance.

Arguments:

writer: AbstractWriter instance to be used for opening file.

path: path to file.

Keyword arguments:

blocks: case's number of blocks.

scenarios: case's number of scenarios.

stages: case's number of stages.

agents: list of element names.

unit: dimension of the elements' data.

is_hourly: if data is hourly. If yes, block dimension will be ignored.

name_length: length of element names.

block_type: case's type of block.

scenarios_type: case's type of scenario.

stage_type: case's type of stage.

initial_stage: stage at which to start registry.

initial_year: year at which to start registry.

allow_unsafe_name_length: allow element names outside safety bounds.

Examples:

Writing and reading a time series into a file

PSRI.open(reader::Type{<:AbstractReader}, path::String; kwargs...)

Method for opening file and reading time series data. Returns updated AbstractReader instance.

Arguments:

reader::Type{<:AbstractReader}: AbstractReader instance to be used for opening file.

path::String: path to file.

Keyword arguments:

is_hourly::Bool: if data to be read is hourly, other than blockly.

stage_type::PSRI.StageType: the PSRI.StageType of the data, defaults to PSRI.STAGE_MONTH.

header::Vector{String}: if file has a header with metadata.

use_header::Bool: if data from header should be retrieved.

first_stage::Dates.Date: stage at which start reading.

verbose_header::Bool: if data from header should be displayed during execution.

Examples:

Writing and reading a time series into a file

PSRClassesInterface.is_hourly — Function

PSRI.is_hourly(ior::AbstractReader)

Returns a Bool indicating whether the data in the file read by PSRI.AbstractReader is hourly.

PSRClassesInterface.max_stages — Function

PSRI.max_stages(ior::AbstractReader)

Returns an Int indicating maximum number of stages in the file read by PSRI.AbstractReader.

PSRClassesInterface.max_scenarios — Function

PSRI.max_scenarios(ior::AbstractReader)

Returns an Int indicating maximum number of scenarios in the file read by PSRI.AbstractReader.

PSRClassesInterface.max_blocks — Function

PSRI.max_blocks(ior::AbstractReader)

Returns an Int indicating maximum number of blocks in the file read by PSRI.AbstractReader.

PSRClassesInterface.max_blocks_current — Function

PSRI.max_blocks_current(ior::AbstractReader)

Returns an Int indicating maximum number of blocks in the cuurent stage in the file read by PSRI.AbstractReader.

PSRClassesInterface.max_blocks_stage — Function

PSRI.max_blocks_stage(ior::AbstractReader, t::Integer)

Returns an Int indicating maximum number of blocks in the stage t in the file read by PSRI.AbstractReader.

PSRClassesInterface.max_agents — Function

PSRI.max_agents(ior::AbstractReader)

Returns an Int indicating maximum number of agents in the file read by PSRI.AbstractReader.

PSRClassesInterface.stage_type — Function

stage_type

PSRClassesInterface.initial_stage — Function

PSRI.initial_stage(ior::AbstractReader)

Returns an Int indicating the initial stage in the file read by PSRI.AbstractReader.

PSRClassesInterface.initial_year — Function

PSRI.initial_year(ior::AbstractReader)

Returns an Int indicating the initial year in the file read by PSRI.AbstractReader.

PSRClassesInterface.data_unit — Function

PSRI.data_unit(ior::AbstractReader)

Returns a String indicating the unit of the data in the file read by PSRI.AbstractReader.

PSRClassesInterface.current_stage — Function

PSRI.current_stage(ior::AbstractReader)

Returns an Int indicating the current stage in the stream of the PSRI.AbstractReader.

PSRClassesInterface.current_scenario — Function

PSRI.current_scenario(ior::AbstractReader)

Returns an Int indicating the current scenarios in the stream of the PSRI.AbstractReader.

PSRClassesInterface.current_block — Function

PSRI.current_block(ior::AbstractReader)

Returns an Int indicating the current block in the stream of the PSRI.AbstractReader.

PSRClassesInterface.agent_names — Function

PSRI.agent_names(ior::AbstractReader)

Returns a Vector{String} with the agent names in the file read by PSRI.AbstractReader.

PSRClassesInterface.goto — Function

PSRI.goto(
    ior::AbstractReader, 
    t::Integer, 
    s::Integer = 1, 
    b::Integer = 1
)

Goes to the registry of the stage t, scenario s and block b.

PSRClassesInterface.next_registry — Function

PSRI.next_registry(ior::AbstractReader)

Goes to the next registry on the PSRI.AbstractReader.

PSRClassesInterface.close — Function

PSRI.close(ior::AbstractReader)

Closes the PSRI.AbstractReader instance.

PSRI.close(iow::AbstractWriter)

Closes the PSRI.AbstractWriter instance.

PSRClassesInterface.convert_file — Function

convert_file

Missing docstring.

Missing docstring for PSRClassesInterface.convert. Check Documenter's build log for details.

PSRClassesInterface.add_reader! — Function

add_reader!

PSRClassesInterface.write_registry — Function

PSRI.write_registry(
    iow::AbstractWriter,
    data::Vector{T},
    stage::Integer,
    scenario::Integer = 1,
    block::Integer = 1,
) where T <: Real

Writes a data row into opened file through PSRI.AbstractWriter instance.

Arguments:

iow: PSRI.AbstractWriter instance to be used for accessing file.

data: elements data to be written.

stage: stage of the data to be written.

scenario: scenarios of the data to be written.

block: block of the data to be written.

PSRClassesInterface.array_to_file — Function

PSRI.array_to_file