Manual

The PSRClassesInterface module provides interfaces to access data structured by PSR to be used in its models. Currently there are two main interfaces.

The interface for studies. This interface is designed to read parameters from the files, some examples are deficit costs, fuel costs, currency, storage capacity etc.

The interface for reading and writing time series data. Time series data in the context of most studies have 4 dimensions (agents, stages, scenarios and blocks). Since studies of renewables with multiple agents, scenarios and stages can get quite big, we have designed different formats that are optimized to some objective (human readability, size, fast reading and writing, etc.).

Both interfaces are defined as a set of methods that need to be implemented to make a different file format work. In this manual we will describe the abstract methods and give concrete examples of code to perform the work needed.

Abstract study interface

PSRClassesInterface.AbstractStudyInterface — Type

AbstractStudyInterface

PSRClassesInterface.initialize_study — Function

initialize_study

PSRClassesInterface.get_vector — Function

get_vector

PSRClassesInterface.max_elements — Function

max_elements

PSRClassesInterface.get_map — Function

get_map

PSRClassesInterface.get_parms — Function

get_parms

PSRClassesInterface.get_code — Function

get_code

PSRClassesInterface.get_name — Function

get_name

PSRClassesInterface.mapped_vector — Function

mapped_vector

PSRClassesInterface.go_to_stage — Function

go_to_stage

PSRClassesInterface.go_to_dimension — Function

go_to_dimension

PSRClassesInterface.update_vectors! — Function

update_vectors!

PSRClassesInterface.description — Function

description

PSRClassesInterface.total_stages — Function

total_stages

PSRClassesInterface.total_scenarios — Function

total_scenarios

PSRClassesInterface.total_blocks — Function

total_blocks

PSRClassesInterface.total_openings — Function

total_openings

PSRClassesInterface.total_stages_per_year — Function

total_stages_per_year

PSRClassesInterface.get_complex_map — Function

get_complex_map

PSRClassesInterface.stage_duration — Function

stage_duration

PSRClassesInterface.stage_block_duration — Function

stage_block_duration

PSRClassesInterface.get_nonempty_vector — Function

get_nonempty_vector

Abstract reader and writer interface

PSRClassesInterface.AbstractReader — Type

AbstractReader

PSRClassesInterface.AbstractWriter — Type

AbstractWriter

PSRClassesInterface.AbstractReaderMapper — Type

AbstractReaderMapper

PSRClassesInterface.AbstractFileType — Type

AbstractFileType

PSRClassesInterface.file_to_array — Function

file_to_array

PSRClassesInterface.file_to_array_and_header — Function

file_to_array_and_header

PSRClassesInterface.open — Function

open

PSRClassesInterface.is_hourly — Function

is_hourly

PSRClassesInterface.max_stages — Function

max_stages

PSRClassesInterface.max_scenarios — Function

max_scenarios

PSRClassesInterface.max_blocks — Function

max_blocks

PSRClassesInterface.max_blocks_current — Function

max_blocks_current

PSRClassesInterface.max_blocks_stage — Function

max_blocks_stage

PSRClassesInterface.max_agents — Function

max_agents

PSRClassesInterface.stage_type — Function

stage_type

PSRClassesInterface.initial_stage — Function

initial_stage

PSRClassesInterface.initial_year — Function

initial_year

PSRClassesInterface.data_unit — Function

data_unit

PSRClassesInterface.current_stage — Function

current_stage

PSRClassesInterface.current_scenario — Function

current_scenario

PSRClassesInterface.current_block — Function

current_block

PSRClassesInterface.agent_names — Function

agent_names

PSRClassesInterface.goto — Function

goto

PSRClassesInterface.next_registry — Function

next_registry

PSRClassesInterface.close — Function

close

PSRClassesInterface.convert_file — Function

convert_file

PSRClassesInterface.convert — Function

convert

PSRClassesInterface.add_reader! — Function

add_reader!

PSRClassesInterface.write_registry — Function

write_registry

PSRClassesInterface.array_to_file — Function

array_to_file